requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
google-api-python-client>=2.152.0
google-auth-oauthlib>=1.2.1
pydantic>=2.7.0
